requirements.txt
streamlit==0.80.0
numpy==1.19.5
tensorflow==2.5.0rc0
keras==2.4.3
opencv-python==4.5.1.48
pillow==8.1.0
